无需经过Vuex访问本地存储, 可以在util.db.js直接访问
